# [stdlib] Optimize lowercasing in `atof` by a lot