Optimize lowercasing in `atof` by a lot.

Improvement:

|name | markdown percentage | magnitude|
|-- | -- | --|
|atof/canada | 0.874 | 7.94|
|atof/mesh | 0.790 | 4.75|


Before:

|name | met (ms) | iters | throughput (GElems/s) | DataMovement (GB/s)|
|-- | -- | -- | -- | --|
|----------- | ----------------- | ----- | --------------------- | --------------------|
|atof/canada | 246.068 | 11 | 0.000452 | 0.008240 |
|atof/mesh | 69.738 | 17 | 0.001047 | 0.008059 |

After:

|name | met (ms) | iters | throughput (GElems/s) | DataMovement (GB/s)|
|-- | -- | -- | -- | --|
|----------- | ------------------ | ----- | --------------------- | -------------------|
|atof/canada | 30.971 | 38 | 0.003588 | 0.065469|
|atof/mesh | 14.678 | 82 | 0.004975 | 0.038293|


CC: @gabrieldemarmiesse this is just helping your implementation shine :)