@soraros 2025-10-04
@martinvuyk Can we introduce a `StringSlice.is_[StringLiteral]()` that implements this unrolling programmatically?